requirements.txt
# Core API framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
starlette==0.27.0

# Database
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.0.1
alembic==1.16.2

# Authentication and Security (ADDED)
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# Data processing
pandas==2.1.3
numpy==1.26.4
openpyxl==3.1.2

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml>=4.9.3

# Validation and utilities
pydantic==2.5.0
pydantic-core==2.14.1

# Configuration and environment
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML>=6.0

# Development and testing (optional)
pytest==7.2.2

# Security
cryptography>=41.0.0

# Other utilities
click==8.2.1
packaging==23.0
certifi==2024.8.30
charset-normalizer==3.1.0
idna==3.4
urllib3>=1.26.15
h11==0.14.0
anyio==3.7.1
sniffio==1.3.0
typing_extensions==4.12.2